refactor: vo 폴더 생성
